add db setup
