+++ curia-back-end/db/setup.js
@@ -1,0 +1,8 @@
+import { connectToDatabase, mongoDb } from "./connection.js";
+const setup = async () => {
+  await connectToDatabase();
+  await mongoDb.db?.dropCollection("users");
+  mongoDb.client.close();
+};
+
+void setup();
