+++ curia-back-end/eslint.config.mjs
@@ -51,4 +51,5 @@ export default tseslint.config(
       ],
     },
   },
+  { files: ["**/*.js"], extends: [tseslint.configs.disableTypeChecked] },
 );
