+++ curia-back-end/package.json
@@ -18,7 +18,9 @@
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
-    "pre-commit": "eslint . && jest"
+    "pre-commit": "eslint . && jest",
+    "setup-db:dev": "NODE_ENV=development node ./db/setup.js",
+    "setup-db:prod": "NODE_ENV=production node ./db/setup.js"
   },
   "dependencies": {
     "@nestjs/common": "^11.0.1",
